Create Final-USF-Quiz
This is a Python class for business students in the University of San Francisco. The program is a quiz that asks the users two questions about USF. If the users were to fail in answering both questions correctly, they have the option to restart the quiz again until they finally answer both questions correctly.